v2/requirements.txt
altgraph==0.17.4
certifi==2025.7.9
charset-normalizer==3.4.2
idna==3.10
packaging==25.0
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
python-docx==1.2.0
requests==2.32.4
setuptools==80.9.0
typing_extensions==4.14.1
urllib3==2.5.0
xmlschema==3.4.4
